legend tooltip rational html
I changed the hover effect of the pie chart and added the tooltip to show the total profit of each product. Additionally I changed the color order of the legend so that each time a new state is clicked, the order of the legend won't change. Finally I added the writeup region but the text is meaningless. (should be added later)